# backend/tests/data/openapi.json
{
  "components": {
    "schemas": {
...
      "CircleVoteSubmit": {
        "properties": {
          "character_id": {
            "title": "Character Id",
            "type": "integer"
          },
          "circle_id": {
            "title": "Circle Id",
            "type": "integer"
          },
          "value": {
            "title": "Value",
            "type": "string"
          },
          "vote_type": {
            "title": "Vote Type",
            "type": "string"
          }
        },
        "required": [
          "circle_id",
          "character_id",
          "vote_type",
          "value"
        ],
        "title": "CircleVoteSubmit",
        "type": "object"
      },
...
      "FinalizeRosterRequest": {
        "properties": {
          "campaign_id": {
            "title": "Campaign Id",
            "type": "integer"
          },
          "circle_id": {
            "title": "Circle Id",
            "type": "integer"
          }
        },
        "required": [
          "campaign_id",
          "circle_id"
        ],
        "title": "FinalizeRosterRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RelationshipPropose": {
        "properties": {
          "circle_id": {
            "title": "Circle Id",
            "type": "integer"
          },
          "from_character_id": {
            "title": "From Character Id",
            "type": "integer"
          },
          "lore": {
            "default": "",
            "title": "Lore",
            "type": "string"
          },
          "rel_type": {
            "title": "Rel Type",
            "type": "string"
          },
          "to_character_id": {
            "title": "To Character Id",
            "type": "integer"
          }
        },
        "required": [
          "circle_id",
          "from_character_id",
          "to_character_id",
          "rel_type"
        ],
        "title": "RelationshipPropose",
        "type": "object"
      },
      "RelationshipRespond": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "counter_lore": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counter Lore"
          },
          "counter_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counter Type"
          },
          "relationship_id": {
            "title": "Relationship Id",
            "type": "integer"
          }
        },
        "required": [
          "relationship_id",
          "action"
        ],
        "title": "RelationshipRespond",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/campaign/finalize-roster": {
      "post": {
        "operationId": "finalize_roster_campaign_finalize_roster_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalizeRosterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finalize Roster"
      }
    },
...
    "/campaign/{campaign_id}/circle-creation-state": {
      "get": {
        "operationId": "get_circle_creation_state_campaign__campaign_id__circle_creation_state_get",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Circle Creation State"
      }
    },
...
    "/circle/relationship/propose": {
      "post": {
        "operationId": "propose_relationship_circle_relationship_propose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationshipPropose"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propose Relationship"
      }
    },
    "/circle/relationship/respond": {
      "post": {
        "operationId": "respond_relationship_circle_relationship_respond_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationshipRespond"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Respond Relationship"
      }
    },
    "/circle/vote": {
      "post": {
        "operationId": "submit_circle_vote_circle_vote_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CircleVoteSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Circle Vote"
      }
    }
  }
}